feat: openapi tensor shapes for all backends
